+++ controllers/user.js
@@ -1,11 +1,17 @@
 const mongoose = require("mongoose");
 const User = require('../models/User.js');
+const Seller = require('../models/Seller.js');
 const bcrypt = require('bcryptjs');
 const { validationResult } = require('express-validator');
 
 const listAllUser =  async (req, res) => {
     try {
-        const users = await User.find().populate([
+        const tenantName = req.tenantName;
+        const tenant = await Seller.findOne({name: tenantName});
+        if (!tenant) {
+            return res.status(400).json({message: "el tenant no es valido"});
+        }
+        const users = await User.find({seller: tenant}).populate([
             {path:'role', select: 'name'},
             {path:'seller', select: 'name description'}
         ]);
@@ -22,13 +28,20 @@ const createUser = async (req, res) => {
         if (!errors.isEmpty()) {
             return res.status(400).json({message: errors.array()})
         }
-        const emailExist = await User.findOne({email: req.body.email});
+        const tenantName = req.tenantName;
+        const tenant = await Seller.findOne({name: tenantName});
+        if (!tenant) {
+            return res.status(400).json({message: "el tenant no es valido"});
+        }
+        const emailExist = await User.findOne({email: req.body.email}).populate([
+            {path:'seller', select: 'name description'},
+        ]);
         if (emailExist) {
             return res.status(400).json({
-                msg: `ya existe un usuario con el email: ${req.body.user_id}`
+                message: `ya existe un usuario con el email: ${req.body.email}, reporatdo en la empresa ${emailExist.seller.name}`
             })
         }
-        const {user_id, name, email, password, role, seller} = req.body;
+        const {user_id, name, email, password, role} = req.body;
         const salt = await bcrypt.genSalt(10);
         const passwordEnc = await bcrypt.hash(password, salt);
 
@@ -38,7 +51,7 @@ const createUser = async (req, res) => {
         user.email = email;
         user.password = passwordEnc;
         user.role = role;
-        user.seller = seller;
+        user.seller = tenant._id;
         user.createDate = new Date();
         user.updateDate = new Date();
         const userSaved = await user.save();
@@ -56,6 +69,11 @@ const getOneUser = async (req, res) => {
         if (!isIdValid) {
             return res.status(400).send('la extensión o formato del id ingresado es invalido');
         };
+        const tenantName = req.tenantName;
+        const tenant = await Seller.findOne({name: tenantName});
+        if (!tenant) {
+            return res.status(400).json({message: "el tenant no es valido"});
+        }
         let user = await User.findById(req.params.userId).populate([
             {path:'role', select: 'name'},
             {path:'seller', select: 'name description'}
@@ -63,6 +81,9 @@ const getOneUser = async (req, res) => {
         if (!user) {
             return res.status(400).send('el usuario a consultar no existe');
         };
+        if (user.seller.name !== tenant.name) {
+            return res.status(400).json({message: "el tenant ingresado no corresponde con el tenant a consultar", user: user.seller._id, tenant: tenant._id});
+        }
         res.status(200).json(user);
     } catch (error) {
         console.log(error)
@@ -76,11 +97,16 @@ const deleteUser = async (req, res) => {
         if (!isIdValid) {
             return res.status(400).send('la extensión o formato del id ingresado es invalido');
         };
+        const tenantName = req.tenantName;
+        const tenant = await Seller.findOne({name: tenantName});
+        if (!tenant) {
+            return res.status(400).json({message: "el tenant no es valido"});
+        };
         let user = await User.findById(req.params.userId);
         if (!user) {
             return res.status(400).send('el usuario a eliminar no existe');
         };
-        const deletedUser = await User.deleteOne({_id: req.params.userId})
+        const deletedUser = await User.deleteOne({_id: req.params.userId, seller: tenant})
         res.status(200).json({resume: deletedUser, user: user});
     } catch (error) {
         console.log(error)
@@ -98,21 +124,25 @@ const updateUser = async (req, res) => {
         if (!isIdValid) {
             return res.status(400).send('la extensión o formato del id ingresado es invalido');
         };
-        let user = await User.findById(req.params.userId);
+        const tenantName = req.tenantName;
+        const tenant = await Seller.findOne({name: tenantName});
+        if (!tenant) {
+            return res.status(400).json({message: "el tenant no es valido"});
+        };
+        let user = await User.findOne({_id: req.params.userId, seller: tenant});
         if (!user) {
             return res.status(400).send('el usuario a actualizar no existe');
         };
-        const userExist = await User.findOne({email: req.body.email, _id:{ $ne: user._id}});
+        const userExist = await User.findOne({_id:{$ne: user._id}, email: req.body.email});
         if(userExist){
             return res.status(400).send('el email ingresado ya está asignado a otro usuario distinto al que está intentando actualizar')
         }
-        const {user_id, name, email, password, role, seller} = req.body;
+        const {user_id, name, email, password, role} = req.body;
         user.userId = user_id;
         user.name = name;
         user.email = email;
         user.password = password;
         user.role = role;
-        user.seller = seller;
         user.updateDate = Date.now();
 
         user = await user.save()
